feat(api): Delete recipe

+++ api/src/main/java/com/a_table/repository/RecipeRepository.java
@@ -3,6 +3,6 @@ package com.a_table.repository;
 import com.a_table.model.RecipeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecipeRepository extends JpaRepository<RecipeEntity, Integer> {
+public interface RecipeRepository extends JpaRepository<RecipeEntity, Long> {
 
 }
